Visual update: fonts and assets
Decided to go for a receipt design. Added a font and some huge  stock photo background images, while making the date navigation a subtle shade.
 Sketched the button assets by myself.
 Decided to change the category color to affect the text and underline colors instead of the entire div, since it is less of an eyesore. Also contemplating whether I want to make the color options fixed instead of a choice from a color picker.

Next up the look of the input prompt and making the gradient into half a radial intead of a full radial since the full one cuts off at the bottom. The button sketches are also going to need some refinement and the entire receipt area needs more ascii art for the proper receipt aesthetic.