Visual and function update
Realised I had been using global variables the entire time, going to fix that soon. Root of this mistake was thinking that Js hadnles variables similarly to Python.

Did some more button designs and decided on some more refined looking ones.

Added my first media query every. Makes the nav buttons take up the sides of the screen on smaller screens. Current width limit is temporary.

Added a basic stats section which at this point sums up and displays the amount of money spent in a day. This section will always be absolutely positioned at the bottom of the content section.

Currently thinking about how to ease in the load times for the big jpegs I use for my main div backgrounds.

Changed the  navbar gradient to a linear one, because it eases into the site better. On a smaller screen there are also gradients to the side implicating the movement with the buttons.