Window prompt layout overhaul
|| BASIC INPUT PROMPT SECTION ||
Made each label, input element pair into their own sections, them sharing space with eachother by flex method.

Added a lot of classes

Changed the display of the buttons to be absolute and moved them into their appropriate positions. At first I messed around with a separate div extending over the current container div until I realised my mistake and misunderstanding of absolute positioning and deleted that div. Also tried working with flex to position the buttons at the bottom. Ultimately decided to just absolutely position them and  have the main two buttons both take 50% of the bottom's width also conveniently centering the text within them.

Added padding and margins to elements aligning them properly

Added placeholder text to the short description text input field

Centered all titles in the prompt window

|| CONFIG SECTION ||

Mostly the same as above, with the exception that I started future proofing it by making the category area it's own section. For future practice each section shall be it's own feature configuration.

Contemplated wether I want to style the hr tag or omit it and get an alternative custom separator solution.

|| Misc. ||

Fixed more global variables
Change the '(optional)' bit of the label to be in the text input's value and added a statement to ignore the value if a user were not to change it. Later found out about the placeholder attribute and used that. Seeing a warning about said new attribute, decided to keep the if statement in place in case it causes trouble and I go back to having the text in the input be value.

|| CSS ||
Added a classes for all new divs played with positioning and sizing.
Also tried using the same paper image I used for the content div and was pleasantly suprised that it didn't look that bad. I'm going to change it in the future, but for now It's better than the previous placeholder solid color.