Calendar View added and implemented
Added calendar view which is accessed by clicking on the date in basic view.

This allows you to skip to any date in the current year.

Dates with transactions are colored with that date's dominant transaction category color.

The date system now runs on a session storage variable called "setDate" and other functions now no longer look to the header text for the current date, but instead receive that from the variable as well. Had to throw some functions around and fix some odd global variable positions during the implementation of this.